+++ src/app/(app)/(docs)/blog/[slug]/page.tsx
@@ -6,11 +6,11 @@ import Link from "next/link";
 import { notFound } from "next/navigation";
 import type { BlogPosting as PageSchema, WithContext } from "schema-dts";
 
-import { LLMCopyButtonWithViewOptions } from "@/components/ai/page-actions";
 import { InlineTOC } from "@/components/inline-toc";
 import { MDX } from "@/components/mdx";
 import { PostKeyboardShortcuts } from "@/components/post-keyboard-shortcuts";
-import { ShareMenu } from "@/components/share-menu";
+import { LLMCopyButtonWithViewOptions } from "@/components/post-page-actions";
+import { PostShareMenu } from "@/components/post-share-menu";
 import { Button } from "@/components/ui/button";
 import { Prose } from "@/components/ui/typography";
 import { SITE_INFO } from "@/config/site";
@@ -132,12 +132,12 @@ export default async function Page({
         </Button>
 
         <div className="flex items-center gap-2">
-          <ShareMenu url={getPostUrl(post)} />
-
           <LLMCopyButtonWithViewOptions
             markdownUrl={`${getPostUrl(post)}.mdx`}
             isComponent={post.metadata.category === "components"}
           />
+
+          <PostShareMenu url={getPostUrl(post)} />
 
           {previous && (
             <Button variant="secondary" size="icon:sm" asChild>
